# Reduced text layout allocations