--- a/src/Avalonia.Base/Media/TextFormatting/FormattingObjectPool.cs
+++ b/src/Avalonia.Base/Media/TextFormatting/FormattingObjectPool.cs
@@ -80,7 +80,7 @@ namespace Avalonia.Media.TextFormatting
                 }
 
                 --_pendingReturnCount;
-                rentedList.Clear();
+                FormattingBufferHelper.ClearThenResetIfTooLarge(rentedList);
 
                 if (_size < MaxSize)
                 {