--- a/src/Avalonia.Base/Media/TextFormatting/SplitResult.cs
+++ b/src/Avalonia.Base/Media/TextFormatting/SplitResult.cs
@@ -26,5 +26,16 @@
         /// The second part.
         /// </value>
         public T? Second { get; }
+
+        /// <summary>
+        /// Deconstructs the split results into its components.
+        /// </summary>
+        /// <param name="first">On return, contains the first part.</param>
+        /// <param name="second">On return, contains the second part.</param>
+        public void Deconstruct(out T first, out T? second)
+        {
+            first = First;
+            second = Second;
+        }
     }
 }