… src/Avalonia.Base/Media/GlyphRun.cs
… src/Avalonia.Base/Media/TextFormatting/TextCharacters.cs
… src/Avalonia.Base/Media/TextFormatting/TextCollapsingProperties.cs
… src/Avalonia.Base/Media/TextFormatting/TextEllipsisHelper.cs
… src/Avalonia.Base/Media/TextFormatting/TextLayout.cs
… src/Avalonia.Base/Media/TextFormatting/TextLeadingPrefixCharacterEllipsis.cs
… src/Avalonia.Base/Media/TextFormatting/TextLineImpl.cs
… src/Avalonia.Base/Media/TextFormatting/TextRunProperties.cs
… src/Avalonia.Base/Media/TextFormatting/TextTrailingCharacterEllipsis.cs
… src/Avalonia.Base/Media/TextFormatting/TextTrailingWordEllipsis.cs
… src/Avalonia.Base/Media/TextFormatting/Unicode/BiDiAlgorithm.cs
… src/Avalonia.Base/Media/TextFormatting/Unicode/BiDiData.cs
… src/Avalonia.Base/Platform/IPlatformRenderInterface.cs
… src/Avalonia.Base/Rendering/Composition/Server/FpsCounter.cs
… src/Avalonia.Base/Utilities/ArrayBuilder.cs
… src/Avalonia.Base/Utilities/ArraySlice.cs
… src/Avalonia.Base/Utilities/BidiDictionary.cs
… src/Avalonia.Base/Utilities/BinarySearchExtension.cs
… src/Avalonia.Controls/TextBoxTextInputMethodClient.cs
… src/Avalonia.Headless/HeadlessPlatformRenderInterface.cs
… src/Skia/Avalonia.Skia/PlatformRenderInterface.cs
… src/Skia/Avalonia.Skia/TextShaperImpl.cs
… src/Windows/Avalonia.Direct2D1/Direct2D1Platform.cs
… src/Windows/Avalonia.Direct2D1/Media/TextShaperImpl.cs
… tests/Avalonia.Base.UnitTests/Media/GlyphRunTests.cs
… tests/Avalonia.Base.UnitTests/Media/TextFormatting/BiDiClassTests.cs
… tests/Avalonia.Base.UnitTests/Media/TextFormatting/FormattingBufferHelperTests.cs
… tests/Avalonia.Base.UnitTests/VisualTree/MockRenderInterface.cs
… tests/Avalonia.Benchmarks/NullRenderingPlatform.cs
… tests/Avalonia.Benchmarks/Text/HugeTextLayout.cs