## What does the pull request do?

This PR removes most temporary allocations during text layout.
Allocations are roughly cut in half in most cases, which gives a measurable performance gain thanks to less garbage collections, especially when layouting multiple small texts.

Details:

- `TextRunProperties.WithTypeface` returns the same instance if the typeface hasn't changed (common case).
- `BidiAlgorithm` is a class again and is reused through a thread static instance (this is how the class was originally designed, as indicated by its comment and https://github.com/SixLabors/Fonts/blob/570875b2664743067e5edd7ac4b69bb789b6c5b4/src/SixLabors.Fonts/Unicode/BidiAlgorithm.cs#L188). This has been measured to be much faster than renting/returning all the arrays it contains to the `ArrayPool` each time.
- `BidiData` is now reusable, like `BidiAlgorithm`.
- `TextFormatting.BidiReorder` is moved to its own class and is reusable. `OrderedBidiRun` and `BidiRange` are now structs.
- `GlyphRun` takes a list of `GlyphInfo` directly instead of 4 lists of indices/advances/offsets/clusters.
   - `ShapedBuffer` was already creating 4 struct wrappers around a `GlyphInfo` list, but this caused extra allocations through boxing.
   - A constructor taking only the glyph indices is kept and gets the glyph advances automatically from the typeface.
   - The Skia and Direct2D1 renderers are simplified to keep only one code path since they are now always provided the advances.
     This causes a slight change of behavior for the Skia renderer. Before, if glyph advances weren't available, the renderer was taking the glyph advances directly from Skia instead of the typeface. Note that in practice this code path was only hit in unit tests, since `TextShaperImpl` always provides glyph information in a standard Avalonia application. Direct2D1 was already using the typeface directly: the two renderers now exhibit the same behavior.
- A `FormattingObjectPool` class is added to pool the many temporary lists used by text layout. Like `BidiAlgorithm`, there's one instance per thread. Pooled usages are fully internal and pooled lists are never stored.
- `TextCollapsingProperties.Collapse` returns an array instead of a list to avoid keeping an extra list around or extracting an array from it.

## Perf changes:

Current master:

|                Method | Wrapping | Trimming |        Mean |    Error |   StdDev |    Gen 0 |   Gen 1 | Allocated |
|---------------------- |--------- |--------- |------------:|---------:|---------:|---------:|--------:|----------:|
|       BuildTextLayout |    False |    False |    356.2 us |  0.88 us |  0.78 us |        - |       - |     22 KB |
| BuildEmojisTextLayout |    False |    False |    775.2 us |  1.30 us |  1.08 us |        - |       - |     27 KB |
|   BuildManySmallTexts |    False |    False |  3,246.5 us | 22.04 us | 20.62 us | 140.6250 | 62.5000 |  2,391 KB |
|  VirtualizeTextBlocks |    False |    False |  4,235.8 us | 24.08 us | 22.52 us | 125.0000 | 15.6250 |  2,273 KB |
|       BuildTextLayout |    False |     True |    402.8 us |  2.16 us |  1.91 us |   7.8125 |  3.9063 |    141 KB |
| BuildEmojisTextLayout |    False |     True |    829.4 us | 12.34 us | 10.94 us |  11.7188 |  3.9063 |    242 KB |
|   BuildManySmallTexts |    False |     True |  3,808.7 us | 20.84 us | 19.50 us | 171.8750 | 78.1250 |  2,814 KB |
|  VirtualizeTextBlocks |    False |     True |  4,265.7 us | 19.00 us | 17.78 us | 125.0000 | 15.6250 |  2,273 KB |
|       BuildTextLayout |     True |    False |  2,283.1 us | 16.33 us | 15.28 us |  23.4375 |  7.8125 |    506 KB |
| BuildEmojisTextLayout |     True |    False | 11,064.5 us | 48.01 us | 42.56 us |  71.4286 |       - |  1,618 KB |
|   BuildManySmallTexts |     True |    False |  4,414.5 us | 10.26 us |  8.57 us | 187.5000 | 93.7500 |  3,197 KB |
|  VirtualizeTextBlocks |     True |    False |  4,372.8 us | 15.51 us | 13.75 us | 125.0000 | 15.6250 |  2,273 KB |
|       BuildTextLayout |     True |     True |  2,387.0 us | 34.01 us | 26.55 us |  31.2500 | 15.6250 |    556 KB |
| BuildEmojisTextLayout |     True |     True | 11,278.1 us | 44.16 us | 41.30 us |  71.4286 |       - |  1,622 KB |
|   BuildManySmallTexts |     True |     True |  4,895.1 us | 29.59 us | 26.23 us | 187.5000 | 93.7500 |  3,555 KB |
|  VirtualizeTextBlocks |     True |     True |  4,374.7 us | 21.03 us | 19.67 us | 125.0000 | 15.6250 |  2,273 KB |



This PR:

|                Method | Wrapping | Trimming |        Mean |    Error |   StdDev |   Gen 0  |   Gen 1 | Allocated |
|---------------------- |--------- |--------- |------------:|---------:|---------:|---------:|--------:|----------:|
|       BuildTextLayout |    False |    False |    332.8 us |  0.40 us |  0.31 us |        - |       - |      9 KB |
| BuildEmojisTextLayout |    False |    False |    695.1 us |  2.08 us |  1.84 us |        - |       - |     10 KB |
|   BuildManySmallTexts |    False |    False |  2,323.0 us | 10.05 us |  8.40 us |  62.5000 | 31.2500 |  1,055 KB |
|  VirtualizeTextBlocks |    False |    False |  3,534.2 us | 14.94 us | 13.25 us |  62.5000 |       - |  1,092 KB |
|       BuildTextLayout |    False |     True |    347.2 us |  1.21 us |  1.13 us |   9.7656 |  3.9063 |    192 KB |
| BuildEmojisTextLayout |    False |     True |    726.4 us |  1.56 us |  1.31 us |  31.2500 | 15.6250 |    545 KB |
|   BuildManySmallTexts |    False |     True |  2,566.1 us |  9.51 us |  8.43 us |  78.1250 | 31.2500 |  1,303 KB |
|  VirtualizeTextBlocks |    False |     True |  3,510.1 us |  9.12 us |  8.09 us |  62.5000 |       - |  1,092 KB |
|       BuildTextLayout |     True |    False |  2,184.2 us | 11.78 us | 10.44 us |  23.4375 |  7.8125 |    471 KB |
| BuildEmojisTextLayout |     True |    False | 11,065.0 us | 68.70 us | 57.36 us |  71.4286 |       - |  1,510 KB |
|   BuildManySmallTexts |     True |    False |  2,941.4 us | 19.35 us | 18.10 us |  78.1250 | 31.2500 |  1,481 KB |
|  VirtualizeTextBlocks |     True |    False |  3,482.9 us | 25.69 us | 24.03 us |  62.5000 |       - |  1,092 KB |
|       BuildTextLayout |     True |     True |  2,222.8 us |  7.80 us |  6.91 us |  23.4375 |  7.8125 |    499 KB |
| BuildEmojisTextLayout |     True |     True | 11,293.8 us | 76.19 us | 71.27 us |  71.4286 |       - |  1,512 KB |
|   BuildManySmallTexts |     True |     True |  3,207.6 us | 18.82 us | 16.68 us |  93.7500 | 46.8750 |  1,721 KB |
|  VirtualizeTextBlocks |     True |     True |  3,538.7 us |  8.45 us |  7.49 us |  62.5000 |       - |  1,092 KB |