@MrJul 2023-01-20
Fixed the buffer reset for stack/dictionary with added tests.
I couldn't get those to trigger in a test app anyway (you'd need a specially crafted text with a lot of isolate pairs).