@Mikolaytis 2023-01-22
I've experianced pooling issues from exceptions. Maybe it's not possible now in avalonia code, but I thing that for safety reasons we should cover all rents with try/finally returns.