@Gillibald 2023-01-22
I saw some of these exceptions during unit testing. No idea how this is possible because the object pool is bound to its own thread.