@MrJul 2023-01-22
These exceptions are mostly for debug diagnosis. If a list isn't returned, it won't impact correctness (only performance): a fresh one will be rented next time. That's why I didn't bother with try/finally most of the time, but maybe this should indeed be adjusted.

There are some tests that raise exceptions that are swallowed, for example `MaskedTextBoxTests.Press_Enter_Add_Custom_Newline`: it throws on `GlyphRun.CreateGlyphRunImpl` because there's no `IPlatformRenderInterface` registered. So the rented lists are never returned. That exception is then caught by `AvaloniaPropertyAccessorPlugin.SendCurrentValue` that's why it's not visible and the test passes.