@Mikolaytis 2023-01-22
if in text code will be an exception somewhere -> RenderList's will not be returned to the pool because return calls are not covered with try/finally.